Implemented feedback from paperprototypetesting of GUI.
M    /src/Ankh.UI/EolStylePropertyEditor.cs - Modified text to be displayed in nameCombo to not include the text "svn:"
M    /src/Ankh.UI/KeywordsPropertyEditor.cs - Modified text to be displayed in nameCombo to not include the text "svn:"
M    /src/Ankh.UI/ExecutablePropertyEditor.cs - Modified text to be displayed in nameCombo to not include the text "svn:"
M    /src/Ankh.UI/IgnorePropertyEditor.cs - Modified text to be displayed in nameCombo to not include the text "svn:"
M    /src/Ankh.UI/PropertyEditorDialog.cs - Moved New-button to avoid reset close to add. Renamed New-button to Reset. Renamed Save-button to Add and Ok-button to Save.
M    /src/Ankh.UI/ExternalsPropertyEditor.cs - Modified text to be displayed in nameCombo to not include the text "svn:"
M    /src/Ankh.UI/ViewLogDialog.cs - Checkbox text "single revision" was modified to "Disable To" and moved. FromDatePicker and ToDatePicker were moved to get closer to fromRevisionComboBox and toRevisionComboBox. Log-area (logRichTextBox) was disabled as default (is enabled if GetLog-button is clicked).  Renamed heading to View Log.
M    /src/Ankh.UI/MimeTypePropertyEditor.cs - Modified text to be displayed in nameCombo to not include the text "svn:"