Initializing uninitialized nodes if it is a project item without child nodes (this excludes directories).
Added a new Section in context menu "Ankh/Tools" where tools wich usually are used rather rarely may be placed.
"Create a list of files..." is the only tool for now. It steps through all nodes below the selected one and gets the corresponding filenames and saves them to a file each on a separate line. This file may then be used as input to i.e. winzip command line client.
This is useful for getting i.e. the sources of a solution an its projects while leaving out automatically generated files and files which do not belong to the solution.

* /src/Ankh/Ankh.2003.csproj
	(this) - Added file FileListCommand.cd
* /src/Ankh/Commands/FileListCommand.cs
	(this) - Added implementation of a file list command
* /src/Ankh/ISolutionExplorer.cs
	(GetSelectionFileNames) - Extended the SolutionExplorer interface to retrieve a list of filenames (recursively) for the currently selected node(s)
* /src/Ankh/Solution/Explorer.cs
	(GetSelectionFileNames) - Added an implementation for the new interface function
* /src/Ankh/Solution/TreeNode.cs
	(RecursiveGetFileNames) - Recursively get file names for this node and the nodes below
	(GetFileNames) - Returns the file names for this node (empty by default). May be overridden by subclasses
* /src/Ankh/Solution/ProjectItemNode.cs
	(GetFileNames) - Returns the file names for this node.
	(GetResources) - If a project item is still uninitialized and it has no children, initialize it now.
	(FindResources) - Don't find resources for unversionable projects.
* /src/Ankh/Solution/ProjectNode.cs
	(GetFileNames) - Returns the file names for this node.
	(IsVersionable) - Returns wether the project is versionable or not
* /src/Ankh/Solution/SolutionNode.cs
	(GetFileNames) - Returns the file names for this node.
* /src/Ankh.Tests/ContextBase.cs
	(GetFileNames) - Added empty implementation of GetSelectionFileNames