Reverted most of r4396.
Allowing to unlock every SvnItem  is not usefull and only gives you 
exceptions. In all our code we assume our status cache is valid. 
Users should use refresh if they made the cache invalid.

The OnError part will not work. This design does not work on multi select.
(The error is eaten, but all other files are ignored).


* src/Ankh/Commands/UnlockCommand.cs
  (OnUpdate) Only allow unlocking of locked items
  
  (OnExecute) Only show locked items; no need to unlock anything else  
  
  Put the ScheduleSvnStatus in a finally to make sure it also runs on partial
  unlocks